// utils/interface.ts
export interface answersType {
    'project-name': string
    'w-template': 'vite-vue3' | 'vite-all' | 'vite-all-mock'
    'w-ui': 'none' | 'element-plus' | 'tdesign' | 'tdesign-mobile'
    'is-ts': boolean
    'is-sass': boolean
    'is-eslint': boolean
    'is-prettier': boolean
    'is-tailwindcss': boolean
    'is-permission': boolean
    'is-leaflet': boolean
    'is-codemirror': boolean
    'is-ckeditor': boolean
}